fix: Days by weeks
We have an invalid behaviour where `-1` act as if the day is present in
the last week. However the rule MUST BE `the last x of the month`.

- `{-1, :monday}` is now the last monday of the month
- `{-2, :monday}` is now the 2nd last monday of the month

I had to change the order of validation execution to execute `interval`
before `days`. This could lead to some breaking change because the
`monthly` interval will be executed BEFORE any days validation (and this
is the correct behaviour)