fix: days validations
When days validation is used with hour validations, we could face an
infinite loop. The days validation used the origin time. So every time we
switch to another day, the time was reset to the origin time.

But if the origin hour was higher than the rule hour, the hour
validation add 1 more day to match the rule hour.

And this create an infinite loop for the following rule starting at any
hour higher than 10:

```
Every Saturday at 10:00AM
```