fix: Monthly interval
We used the value for every X months to bump the month by the value. 

However if the month is shift by any others validation we continue to shift the same amount +X months. Instead we should use a relative shift.

Let’s say we have `Every 2 months on the first Monday starting on 2025-05-19`

1. I switch to 1st Monday -> 2025-06-02
2. Not every 2 months -> 2025-08-01 (not every 2 months)

Instead it should shift to to `2025-07-01`